+++ pypsa/descriptors.py
@@ -289,6 +289,23 @@ def get_committable_i(n, c):
     return idx.rename(f"{c}-com")
 
 
+def get_fixed_operation_i(n, c):
+    """
+    Getter function.
+
+    Get the index of elements with fixed profiles of a given component.
+
+    Currently only considered for Generator and Link component.
+    """
+    if c not in {"Generator", "Link"}:
+        return pd.Index([])
+    fix_i = n.get_non_extendable_i(c)
+    p_set = get_switchable_as_dense(n, c, "p_set")
+    fixed_operation_i  = p_set.where(p_set != 0).dropna(how='all', axis=1).columns
+    idx = fixed_operation_i.intersection(fix_i)
+    return idx
+
+
 def get_active_assets(n, c, investment_period):
     """
     Getter function.