+++ doc/references/release-notes.rst
@@ -34,6 +34,12 @@ Upcoming Release
 
 * Bugfix: The global constraint on the total transmission costs now includes the weight of the investment periods and persistence of investment costs of active assets in multi-horizon optimisations.
 
+* Performance: Dispatch variables for non-extendable links and generators with
+  fixed dispatch profiles (``p_set``) are now no longer created and
+  added as constants to the optimisation problem.
+
+* The ``p_set`` and ``q_set`` of all components but the Load component now default to NaN rather than zero.
+
 PyPSA 0.28.0 (8th May 2024)
 =================================